Title: 최솟값 3개, Time: 1149ms, Memory: 29MB, Status: Passed - CodeTree

+++ 240917/최솟값 3개/top-3-smallest-number.kt
@@ -18,7 +18,7 @@ fun main() {
             val first = pq.poll()
             val second = pq.poll()
             val third = pq.poll()
-            println(first * second * third)
+            println(first.toLong() * second.toLong() * third.toLong())
             pq.add(first)
             pq.add(second)
             pq.add(third)
